feature activity no banco de dados.

+++ src/main/java/br/edu/ifsp/ifitness/model/Activity.java
@@ -14,7 +14,7 @@ public class Activity implements Serializable {
 	private ActivityType type;
 	private LocalDate date;
 	private Double distance;
-	private Double duration;
+	private Integer duration;
 	private User user;
 
 	public Long getId() {
@@ -49,18 +49,18 @@ public class Activity implements Serializable {
 		this.distance = distance;
 	}
 
-	public Double getDuration() {
+	public Integer getDuration() {
 		return duration;
 	}
 
-	public void setDuration(Double duration) {
+	public void setDuration(Integer duration) {
 		this.duration = duration;
 	}
-	
+
 	public User getUser() {
 		return user;
 	}
-	
+
 	public void setUser(User user) {
 		this.user = user;
 	}
